Disable CSS transitions in tests
CSS transitions often break tests. For example, a popup can take some
time to be fully shown if there is a transition time applied to its
show effect. The test will see that the popup is present and will try to
use it, but the element is ready yet.

This disables the CSS transitions in the test environment and removes
timing issues related to CSS.